% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{eval_expa}
\alias{eval_expa}
\title{Evaluate expansions in the parent environment}
\usage{
eval_expa(expa)
}
\arguments{
\item{expa}{an <code>expansions</code> object}
}
\description{
Evaluate expansions in the parent environment
}
