% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{eval_expa_within}
\alias{eval_expa_within}
\title{Evaluate expansions in the environment of a dataframe, list,
or timeseries object.}
\usage{
eval_expa_within(expa, data)
}
\arguments{
\item{expa}{an \code{\link{expansions}} object}

\item{data}{dataframe, list, \code{\link[stats]{ts}} or
\code{\link[regts]{regts}} within which the expanded expressions are
 executed}
}
\description{
Evaluate expansions in the environment of a dataframe, list,
or timeseries object.
}
