% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_numeric}
\alias{expand_numeric}
\title{Expands a range of integerss into all the integers it contains}
\usage{
expand_numeric(x, sep)
}
\arguments{
\item{x}{A string containing a condensed range of numbers; "a-b".}

\item{sep}{The character separating the integers for output. Defaults to `NULL`,
which will return a vector.}
}
\value{
Either a vector of integers (if `sep == NULL`) or a character string.
}
\description{
Expands a range of integerss into all the integers it contains
}
\examples{
`expand_numeric("1-5")`
`expand_numeric("20-26", sep = " | ")`
}
