% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{expansions}
\alias{expansions}
\title{Expand R expressions with and without aggregation}
\usage{
expansions(x)
}
\arguments{
\item{x}{one or more R-expressions}
}
\description{
Expand R expressions with and without aggregation
}

