% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{eval_expa_within}
\alias{eval_expa_within}
\title{Evaluate expansions in the environment of a dataframe or list}
\usage{
eval_expa_within(code, x)
}
\arguments{
\item{an}{<code>expansions</code> object}

\item{within}{dataframe or list within which the expanded expressions are
executed}
}
\description{
Evaluate expansions in the environment of a dataframe or list
}

