% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_numeric}
\alias{expand_numeric}
\title{Expands a range of integers into all the integers it contains}
\usage{
expand_numeric(x, sep = NULL, replace = FALSE)
}
\arguments{
\item{x}{A character vector containing condensed ranges of numbers;
"a-b" or "a - b".}

\item{sep}{The character separating the integers for output. Defaults to `NULL`,
which will return a list of integer vectors.}

\item{replace}{Controls if the expanded range is extracted into a list (`FALSE`) or replaced
within the strings (`TRUE)`)}
}
\value{
Either a list of character vectors (if `replace == FALSE`) or a character vector
similar to the input vector.
}
\description{
Expands a range of integers into all the integers it contains
}
\examples{
`expand_numeric("Precincts 1-5")`
`expand_numeric("Precincts 1-5,7-9", replace = TRUE)`

}
